refactor: set uniform userHeader name
